Made Separate Interfaces cs
Created new classes for each separate interface.
