+++ Assets/Scripts/Interfaces.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9708aa0ad9175a447a7cc372328a2f4f
+folderAsset: yes
+timeCreated: 1463495686
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
