+++ Assets/Scripts/Interfaces/IAttackable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d2e0275cea7f4014aaddbd01a4c142a7
+timeCreated: 1463496058
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
