+++ Assets/Scripts/Interfaces/IControllable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5396626fd2919124882032db149ae9f9
+timeCreated: 1463495724
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
