+++ Assets/Scripts/Interfaces/IStats.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0ed89b7f1b3901441a4993e1dfb24599
+timeCreated: 1463496776
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
