+++ Assets/Scripts/Interfaces/Parent.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 741447dd672dfec4796e83eeee94830a
+timeCreated: 1463495977
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
