+++ Assets/Scripts/MyCamera.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine;
 using System.Collections;
+using Interfaces;
 using Units;
 using Units.Controller;
