+++ Assets/Scripts/Units/Fortress.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Interfaces;
 using Library;
 using UI;
 using UnityEngine;
