+++ Assets/Scripts/Units/Player.cs
@@ -1,6 +1,6 @@
 ﻿// Unit class used for storing Player Data.
 using System.Collections.Generic;
-
+using Interfaces;
 using UI;
 using UnityEngine;
