+++ Assets/Scripts/Units/Skills/Fireball.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using Interfaces;
+using UnityEngine;
 using Library;
 
 namespace Units.Skills
