+++ Assets/Scripts/Units/Skills/Skill.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-
+using Interfaces;
 using UnityEngine;
 
 using Library;
